+++ frontend/bindings/hermes/models.ts
@@ -20,4 +20,5 @@ export interface Settings {
     "figureAlignment": string;
     "chartWidth": string;
     "paperSize": string;
+    "showOutline": boolean;
 }